front end hooked up with images

+++ webapp/src/main/webapp/ar/edu/itba/paw/webapp/controller/PropertyController.java
@@ -2,6 +2,7 @@ package ar.edu.itba.paw.webapp.controller;
 
 import java.util.List;
 
+import ar.edu.itba.paw.model.Property;
 import ar.edu.itba.paw.webapp.Utilities.UserUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -27,7 +28,7 @@ public class PropertyController {
     }
 
     @RequestMapping(value = "{id}", method = RequestMethod.GET)
-    public ModelAndView get(@PathVariable("id") int id) {
+    public ModelAndView get(@PathVariable("id") long id) {
         final ModelAndView mav = new ModelAndView("detailedProperty");
         mav.addObject("property", propertyService.getPropertyWithRelatedEntities(id));
         return mav;
